Add movie endpoint, test, testing batch file etc.
This change implements the add_movie function for the route
"/movies". The actual creation of the movie is not tested yet,
only that the command completes successfully.
In addition, changes have been made so that testing can
be convenienty run using a batch file, and
a requirements.txt has been added.